+++ Decision/RandomComputerDecision.cpp
@@ -11,10 +11,10 @@ RandomComputerDecision::RandomComputerDecision()
 UnitPtr RandomComputerDecision::chooseUnitToAttack(const Unit& decidingUnit,
 	const Gladiators& arena)const
 {
-	index unitToAttack = randomNumber(arena.size() - 1);
+	index unitToAttack = randomNumber((int)arena.size() - 1);
 	while (arena[unitToAttack]->isAlly(decidingUnit) ||
 		!arena[unitToAttack]->isAlive())
-		unitToAttack = randomNumber(arena.size() - 1);
+		unitToAttack = randomNumber((int)arena.size() - 1);
 	return arena[unitToAttack];
 }
 
@@ -23,18 +23,18 @@ MagicPtr RandomComputerDecision::chooseMagicToCast(const Unit& decidingUnit,
 {
 	if (!decidingUnit.m_magicBook.canCastAnySpell())
 		return nullptr;
-	index magicToCastIndex = randomNumber(decidingUnit.m_magicBook.size() - 1);
+	index magicToCastIndex = randomNumber((int)decidingUnit.m_magicBook.size() - 1);
 	while (!decidingUnit.isEnoughManaFor(decidingUnit.m_magicBook[magicToCastIndex]))
-		magicToCastIndex = randomNumber(decidingUnit.m_magicBook.size() - 1);
+		magicToCastIndex = randomNumber((int)decidingUnit.m_magicBook.size() - 1);
 	return decidingUnit.m_magicBook[magicToCastIndex]->clone();
 }
 
 UnitPtr RandomComputerDecision::chooseUnitToCast(const Unit& decidingUnit, 
 	const MagicPtr& magicToCast, const Gladiators& arena)const
 {
-	index unitToCast = randomNumber(arena.size() - 1);
+	index unitToCast = randomNumber((int)arena.size() - 1);
 	while (isWrongSpellToCast(decidingUnit, *arena[unitToCast], magicToCast))
-		unitToCast = randomNumber(arena.size() - 1);
+		unitToCast = randomNumber((int)arena.size() - 1);
 	return arena[unitToCast];
 }
